User role management with LDAP Groups
Added the following features:

User's roles can be managed by OpenLDAP:
The user's roles are allocated to LDAP Groups.
If a user is member of a group, he obtains the
matching role. The groups to roles allocation
can be configured in the wordpress admin interface.
If the user is not in an allocated group, a
default role is used.

LDAP filter syntax is used to search for user
dn and group dn. This makes the plugin much
more flexible to different LDAP configurations.
The filter can be configured in the admin interface.

After every login, the user details can be synchronized
to the LDAP entries. This feature can be activated
or deactivated in the wp admin interface.