Feature: Search for LDAP account
Rather than constructing the DN of an account, this allows to bind
using a dedicated LDAP account for searching, then searching for
an LDAP account, then doing the actual password check.

This allows for more flexibility since the Wordpress account name
can be stored in any field of the LDAP account. Also allows n:1
mapping of Wordpress to LDAP accounts.